# design/openapi.yaml
openapi: 3.1.0
info:
  title: parkmitra API
  version: 1.0.0
  license:
    name: MIT
    identifier: MIT
  description: |
    Short-stay parking against idle private capacity, Hyderabad.

    Written at phase 3, BEFORE any implementation, and binding for the rest of the
    project (programme rule: "API contract written before any implementation. It is
    binding for the rest of the project.").

    Two rules that shape every path below:

    - **Money is integer paise.** Never rupees, never a float. `4000` is ₹40.00.
      The client never sends an amount; the server computes it. See ADR-003.
    - **A booking is addressed by its `reference_code`, never by its id.** Holding the
      code is the authorisation to read or modify that booking. Ids are internal and
      never appear in a URL. See ADR-002.

# No authentication scheme applies anywhere in v1: there are no accounts
# (ADR-002). The three /bookings/{reference_code} operations are authorised by
# possession of a ~50-bit reference code supplied as a path parameter, which
# OpenAPI cannot express as a securityScheme. Declaring `security: []` here is
# the accurate statement — "no scheme applies" — rather than leaving it
# undefined and letting a reader assume it was forgotten.
security: []
servers:
  - url: /api
    description: Same-origin route handlers
...
paths:
...
  /bookings/{reference_code}/pay:
    post:
      tags: [bookings]
      operationId: payBooking
      summary: Authorise payment through the configured provider
      description: |
        Authorise payment through the configured provider (ADR-005). The default
        provider is `simulated` — no money moves; a `payments` row is recorded and
        the booking moves to `confirmed`. The provider is selected by the
        `PAYMENT_PROVIDER` environment variable.

        A 422 is returned when the provider declines the charge (e.g. insufficient
        funds with a real gateway). The simulated provider never declines.
      parameters:
        - name: reference_code
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{10}$" }
      responses:
        "200":
          description: Confirmed
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409":
          description: Already paid, or the booking is cancelled or expired
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: Payment provider declined the charge
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
              example:
                error: payment_failed
                message: Payment was declined. Please try again.
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /bookings/{reference_code}/verify/start:
    post:
      tags: [verification]
      operationId: startVerification
      summary: Send an OTP to the phone number on this booking
      description: |
        The phone number is read from the booking row on the server side — it is
        NEVER accepted as a request parameter. This prevents the phone-number
        oracle attack: an attacker cannot submit phone numbers to discover which
        ones have bookings (ADR-002, threat model T1).

        The response is deliberately identical whether the booking exists, does
        not exist, or is cancelled. A known and an unknown reference_code produce
        byte-identical 200 bodies. This is the security invariant this endpoint
        must maintain — any difference leaks information.

        Rate limited: 10 requests per minute per IP (scope `verify`).
      parameters:
        - name: reference_code
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{10}$" }
      responses:
        "200":
          description: |
            Always 200 with a generic message. Whether an OTP was actually sent
            is not revealed — that would be the oracle.
          content:
            application/json:
              schema:
                type: object
                required: [message]
                properties:
                  message: { type: string }
              example:
                message: "If this booking exists, a verification code has been sent."
        "429":
          $ref: "#/components/responses/TooManyRequests"
  /bookings/{reference_code}/verify/check:
    post:
      tags: [verification]
      operationId: checkVerification
      summary: Verify a 6-digit OTP code
      description: |
        The OTP is looked up by booking_id (never by phone number). The response
        is identical for unknown bookings and failed verifications — no oracle.

        On success, `bookings.phone_verified` is set to `true`.

        Attempt limiting: at most 5 wrong guesses per OTP. After 5 failures the
        OTP is burned and the driver must request a new one via `/verify/start`.
        The attempt counter is protected against concurrency by a row-level lock
        (`SELECT ... FOR UPDATE`).

        Rate limited: 10 requests per minute per IP (scope `verify`).
      parameters:
        - name: reference_code
          in: path
          required: true
          schema: { type: string, pattern: "^[0-9A-HJKMNP-TV-Z]{10}$" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  pattern: "^\\d{6}$"
                  description: The 6-digit OTP code.
                  example: "482917"
      responses:
        "200":
          description: |
            Always 200. `verified: true` on success; `verified: false` for wrong
            code, expired OTP, max attempts, or unknown booking — all
            indistinguishable.
          content:
            application/json:
              schema:
                type: object
                required: [verified]
                properties:
                  verified: { type: boolean }
        "400":
          description: Missing or malformed `code` field
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
components:
  responses:
...
    NotFound:
      description: No such booking
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    TooManyRequests:
      description: |
        Rate limited. 20 requests per minute per IP on the booking lookup.
        The `Retry-After` header contains the number of seconds until the
        current window resets.
      headers:
        Retry-After:
          description: Seconds until the rate-limit window resets.
          schema: { type: integer, minimum: 1, maximum: 60 }
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: rate_limited
            message: Too many requests. Try again later.
  schemas:
...
    Window:
      type: object
      description: Half-open. `end` is exclusive.
      required: [start, end]
      properties:
        start: { type: string, format: date-time }
        end: { type: string, format: date-time }
...
    Booking:
      type: object
      required:
        [reference_code, status, window, amount_paise, spot_name, bay_label]
      properties:
        reference_code:
          type: string
          example: 7K2M9QX4TB
          description: Treat as a secret. This is the authorisation for this booking.
        status:
          type: string
          enum: [pending, confirmed, cancelled, expired]
        window: { $ref: "#/components/schemas/Window" }
        amount_paise: { type: integer, example: 8000 }
        spot_name: { type: string }
        bay_label: { type: string, example: B-12 }
        address_line: { type: string }
        arrived_at: { type: [string, "null"], format: date-time }
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Stable machine-readable code.
          example: bay_unavailable
        message:
          type: string
          description: |
            Human-readable. Must never contain personal data, SQL, or a stack trace.